requirements.txt
python-telegram-bot==11.1.0
beautifulsoup4==4.7.1
jdatetime==3.2.0
psycopg2==2.8.2